Simulated data in the expected format.
Updated the simulate_data file by simulating the lead concentration of water in 1000 Torontonian house, while also randomly assigning each house a postal code. I have also updated the remainder of the information that was in the original file. The next step is creating a few tests to ensure the simulated data works as expected.